Removed sandbox directive for unsupported browsers
Firefox does not yet support the CSP sandbox directive
(see https://wiki.mozilla.org/Features/Platform/CSP_Sandbox)
and throws an error message if it finds it. For now I've removed
it for all browsers except for IE10+ and Chrome 14+. I wasn't able
to determine when Chrome started to support this feature.
For all unrecognised browsers 'enforcePolicy' will determine whether
the sandbox directive is removed or not.