Updated Code Example
The code example for the Recipe was missing a few things. There was no include for `csp`. I added the include for the contentSecurityPolicy as referenced in the previous example. Also, I made `csp` a little more clear by just expanding it out to be consistent with the first example.

Thanks for all of your hard work. I thought I would contribute a little to make things better for the next folks coming in. Security is a tough thing to get your mind around. Hopefully this helps the next person working on implementing this!